replace Authorisation with Authorization
looks like a lot of schemas / documents are spelling Authorisation with
an `s`. A quick search revealed that this is a cultural british thing:

```
While many people in the UK prefer the 's' (and some will make a fuss about 'z' being "American"), both forms are recognised in the UK, and the Oxford Dictionaries (which are widely taken as authorities) prefer 'z'.
```

since some examples in the docs are mentioning Authorization with `z`
and HTTP RFCs are using `z` aswell, i think it's better for us to do
this refactoring to prevent confusion in the future.

```
4.2.  Authorization

   The "Authorization" header field allows a user agent to authenticate
   itself with an origin server -- usually, but not necessarily, after
   receiving a 401 (Unauthorized) response.  Its value consists of
   credentials containing the authentication information of the user
   agent for the realm of the resource being requested.

     Authorization = credentials

   If a request is authenticated and a realm specified, the same
   credentials are presumed to be valid for all other requests within
   this realm (assuming that the authentication scheme itself does not
   require otherwise, such as credentials that vary according to a
   challenge value or using synchronized clocks).

   A proxy forwarding a request MUST NOT modify any Authorization fields
   in that request.  See Section 3.2 of [RFC7234] for details of and
   requirements pertaining to handling of the Authorization field by
   HTTP caches.
```
https://tools.ietf.org/html/rfc7235#section-4.2